test(vm): migrate DataToken `.Float` reads to `.Double` after number=Double remap
After the number→Double remap, raw numeric literals like `10.0` flow through
`DataToken.op_Implicit__SystemDouble`, so the stored `TokenType` is Double, not
Float. Reading the same token back via `DataToken.Float` (which expects
TokenType.Float) throws the EXTERN — the affected fixtures were not testing
Float specifically, just round-tripping a `number` value through a DataList /
DataDictionary.

Switch 20 fixtures to read with `.Double`, matching the post-remap storage
type. Tests that intentionally exercise UdonFloat / UdonSharp Float storage
keep `.Float` (none of the migrated files use UdonFloat).

VM result: 342/363 → 362/363 (1 remaining failure unrelated to this change).

+++ tests/vm/cases/data_dictionary_trygetvalue_flow.ts
@@ -12,7 +12,7 @@ export class DataDictionaryTrygetvalueFlow extends UdonSharpBehaviour {
     const dict = new DataDictionary();
     dict.SetValue(new DataToken("hp"), new DataToken(120.0));
     const hasHp = dict.ContainsKey(new DataToken("hp"));
-    const hpValue = hasHp ? dict.GetValue(new DataToken("hp")).Float : 0.0;
+    const hpValue = hasHp ? dict.GetValue(new DataToken("hp")).Double : 0.0;
     const hasMp = dict.ContainsKey(new DataToken("mp"));
 
     Debug.Log(hasHp);
